pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pygamepal"
version = "1.0.3"
authors = [
  { name="Rik Cross", email="rik.j.cross@gmail.com" },
]
description = "A library of classes and helper functions to support game development in Pygame"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "pygame-ce",
]

[project.urls]
"Homepage" = "https://github.com/rik-cross/pygamepal"
"Suggestions and bugs" = "https://github.com/rik-cross/pygamepal/issues"

[tool.hatch.build.targets.sdist]
exclude = [
  "/docs",
  "/examples",
  "/templates",
]